+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 * AWS: Create and delete DNS records
 * Remove obsolete `cluster-create` command. Use `cluster create` instead.
+* Added support for setting Azure security groups. See [sye-azure/README.md](sye-azure/README.md) for details.
 
 ## [1.0.0] - 2018-02-23